refactor(LottoNumberGenerator): domain -> service로
