+++ src/main/java/lotto/service/LottoNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.service;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.List;
